\name{RCABLAB-package}
\alias{RCABLAB-package}
\alias{RCABLAB}
\docType{package}
\title{Reads data from the Earth System Datacube}
}
\description{This is the R interface for reading data from the Earth System data cube.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCABLAB}
\packageIndices{RCABLAB}
}
\author{Fabian Gans}

Maintainer: \packageMaintainer{RCABLAB}
}

\keyword{ package }


