\name{RDataCubeReader-package}
\alias{RDataCubeReader-package}
\alias{RDataCubeReader}
\docType{package}
\title{Reads data from the Earth System Datacube}
}
\description{This is the official R interface for reading data from the Earth System data cube.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RDataCubeReader}
\packageIndices{RDataCubeReader}
}
\author{Fabian Gans}

Maintainer: \packageMaintainer{RDataCubeReader}
}

\keyword{ package }


